+++ models/src/anemoi/models/layers/residual.py
@@ -20,7 +20,7 @@ from torch_geometric.data import HeteroData
 from anemoi.models.distributed.graph import gather_channels
 from anemoi.models.distributed.graph import shard_channels
 from anemoi.models.distributed.shapes import apply_shard_shapes
-from anemoi.models.layers.graph_providers import ProjectionGraphProvider
+from anemoi.models.layers.graph_provider import ProjectionGraphProvider
 from anemoi.models.layers.sparse_projector import SparseProjector
