+++ models/src/anemoi/models/models/encoder_processor_decoder.py
@@ -20,7 +20,7 @@ from torch.distributed.distributed_c10d import ProcessGroup
 from anemoi.models.distributed.graph import shard_tensor
 from anemoi.models.distributed.shapes import get_or_apply_shard_shapes
 from anemoi.models.distributed.shapes import get_shard_shapes
-from anemoi.models.layers.graph_providers import create_graph_provider
+from anemoi.models.layers.graph_provider import create_graph_provider
 from anemoi.models.models import BaseGraphModel
 from anemoi.utils.config import DotDict
